requirements.txt
streamlit
torch
torchvision
Pillow
opencv-python-headless
streamlit-webrtc
